LSP: Rename widget config files
