+++ spyder/plugins/completion/languageserver/widgets/snippetsconfig.py
